user-service/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
requests==2.31.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0